+++ packages/api-models/src/types/Errors.ts
@@ -1,0 +1,4 @@
+import { Error } from "./Error"
+export interface Errors {
+    readonly errors: readonly Error[]
+}