+++ apps/api/CHANGELOG.md
@@ -9,11 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   `existing_uuid` parameter for `postUpload` operation.
+-   `409` response for `postUpload` operation.
+
 ### Changed
 
 ### Deprecated
 
 ### Fixed
+
+-   No more empty strings for `Error.field`. (Should be `undefined` instead.)
 
 ### Removed